+++ devtools/replidev/src/podman/pod_start_cmd.rs
@@ -144,6 +144,13 @@ where
             let error = ErrorKind::podman_failed("run");
             return Err(error.into());
         }
+
+        // If the container has a start delay wait a bit.
+        if let Some(delay) = container.start_delay {
+            println!("--> Wating {}s for {} to start", delay, con_name);
+            let delay = std::time::Duration::from_secs(delay);
+            std::thread::sleep(delay);
+        }
     }
     Ok(())
 }